requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg[binary]>=3.1.0
redis==5.0.1
celery==5.3.4
cryptography==41.0.7
python-dotenv==1.0.0
fastapi==0.104.1
uvicorn==0.24.0
django-environ==0.11.2

